Initial code commit. Added message.go and server.go
Most of the work was on the message.go file, where more or less the entire
interface for the messages has been worked out. Messages are defined using one
of three types, and the NewMessage or ParseMessage functions should be used to
create messages correctly.

The server.go file mostly has the basic interfaces that are going to be used.